reference/requirements.txt
python-telegram-bot[job-queue]>=20.0
python-dotenv>=1.0.0
pandas>=2.0.0
pytz>=2023.3
requests>=2.31.0
aiohttp>=3.9.0
nba_api>=1.4.0
